--- src/CarbonAware.NewCLI.UnitTests/CarbonAwareCLIUnitTests.cs
@@ -1,15 +1,0 @@
-namespace CarbonAware.NewCLI.UnitTests;
-
-public class CarbonAwareCLIUnitTests
-{
-    [SetUp]
-    public void Setup()
-    {
-    }
-
-    [Test]
-    public void Test1()
-    {
-        Assert.Pass();
-    }
-}
