+++ src/CarbonAware.NewCLI/CommandKeywords/Emissions/EmissionsForecastCommand.cs
@@ -1,0 +1,19 @@
+﻿using CarbonAware.Aggregators.CarbonAware;
+using System;
+using System.Collections.Generic;
+using System.CommandLine;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace CarbonAware.NewCLI.CommandKeywords
+{
+    internal static class EmissionsForecastCommand
+    {
+        public static void AddEmissionsForecastCommand(this Command parent, ICarbonAwareAggregator aggregator)
+        {
+            throw new NotImplementedException();
+        }
+
+    }
+}
